requirements.txt
flask==3.0.2
python-dotenv==1.0.1
langchain==0.2.16
langchain-openai==0.1.23
langchain-community==0.2.16
chromadb==0.5.5
tiktoken==0.7.0